Gestion de docentes implementado
Aun no se mapean correctamente los historicos y formatos A relacionados al docente

+++ src/main/java/co/edu/unicauca/asae/cleanarquitecture/infraestructura/output/persistencia/entidades/DocenteEntity.java
@@ -20,7 +20,7 @@ public class DocenteEntity {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int idDocente;
+    private Long idDocente;
 
     @Column(name = "nombres_docente", length = 100)
     private String nombresDocente;
@@ -39,4 +39,6 @@ public class DocenteEntity {
 
     @OneToMany(fetch = FetchType.EAGER, mappedBy = "objDocente")
     private Set<FormatoAEntity> formatosA;
+
+
 }
